potential menu class
problems:
Couldnt get pipes to be infinite by moving the pipes already passed by to the end of the pipes. once the bird collides, the pipes are not in sync to the randomize anymmore lmao.
The score would also +=2 for some reason after passing the last pipe while trying the infinite pipes

Score is the same once the bound is hit
then the bird is back to the beggining anf pipes are randmozied again